% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armainv.R
\name{armainv}
\alias{armainv}
\title{Inverse of Matrix Using RcppArmadillo}
\usage{
armainv(x)
}
\arguments{
\item{x}{Matrix to be inverted}
}
\value{
Inverted matrix
}
\description{
Inverse of Matrix Using RcppArmadillo
}
\examples{

set.seed(123)
X <- crossprod(matrix(rnorm(25),5,5))
armainv(X)

}

