% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armainv.R
\name{armainv_pd}
\alias{armainv_pd}
\title{Inverse of Symmetric Positive-definite Matrix Using RcppArmadillo}
\usage{
armainv_pd(x)
}
\arguments{
\item{x}{Matrix to be inverted}
}
\value{
Inverted matrix
}
\description{
Inverse of Symmetric Positive-definite Matrix Using RcppArmadillo
}
\examples{

set.seed(123)
X <- crossprod(matrix(rnorm(25),5,5))
armainv_pd(X)

}

